Fixed namespace issues in the backend, and moved authExceptions into the right directory
