+++ backend/FlashIdBackend/Application/Common/Interfaces/ServiceInterfaces/IOnboardingService.cs
@@ -1,6 +1,6 @@
 using Application.Features.Onboarding.Dtos;
 
-namespace Application.Common.Interfaces;
+namespace Application.Common.Interfaces.ServiceInterfaces;
 
 public interface IOnboardingService
 {
