+++ backend/FlashIdBackend/Application/Common/Services/InstitutionService.cs
@@ -1,4 +1,3 @@
-using Application.Common.Interfaces;
 using Application.Common.Interfaces.RepositoryInterfaces;
 using Application.Common.Interfaces.ServiceInterfaces;
 using Application.Common.Mapping;
