+++ backend/FlashIdBackend/Infrastructure/DependencyInjection.cs
@@ -1,12 +1,8 @@
-using Application.Common.Interfaces;
 using Application.Common.Interfaces.ProviderInterfaces;
 using Application.Common.Interfaces.RepositoryInterfaces;
-using Application.Common.Interfaces.ServiceInterfaces;
-// using Application.Common.Services;
 using Domain.Entities;
 using Infrastructure.Providers;
 using Infrastructure.Repositories;
-using Infrastructure.Services;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -24,8 +20,6 @@ public static class DependencyInjection
         services.AddScoped<IAuthRepository, AuthRepository>();
         services.AddScoped<IMockGovernmentRegistryRepository, MockGovernmentRegistryRepository>();
         services.AddScoped<IOnboardingRepository, OnboardingRepository>();
-
-        // services.AddScoped<IOnboardingService, OnboardingService>();
 
         services.AddScoped<ICitizenRepository, CitizenRepository>();
         services.AddScoped<IInstitutionRepository, InstitutionRepository>();
