+++ backend/FlashIdBackend/Infrastructure/Repositories/CitizenRepository.cs
@@ -7,7 +7,7 @@ using Infrastructure.Data;
 
 using Microsoft.EntityFrameworkCore;
 
-namespace Infrastructure.Services;
+namespace Infrastructure.Repositories;
 
 public class CitizenRepository : ICitizenRepository
 {
